Implemented additional alive check for interface

+++ SimpleHealthCheck_Client_Cronjob.sh
@@ -22,28 +22,31 @@ do
 
     if [ "$SURICATA_PID" != "" ]
     then
-        echo "[D] Suricata PID File Found"
-
-        if ! ps -p $SURICATA_PID > /dev/null 2> /dev/null
+        if ! ps aux | grep suricata | grep $SURICATA_IF_NAME 2> /dev/null
         then
-            echo "[INFO] Suricata is NOT running on interface $SURICATA_IF_NAME! Will try to fix this..."
-            echo "[D] Removing potential stuck PID file"
-            rm "$VARPATH$SURICATA_IF_NAME"*
+            echo "[D] Suricata PID File Found"
 
-            echo "[D] Restarting Suricata on interface"
-            SURICATA_CONFIG_PATH=$(find /usr/local/etc/suricata -path /usr/local/etc/suricata/suricata_*_$SURICATA_IF_NAME)
-            SURICATA_TUN_NUMBER=$(echo $SURICATA_CONFIG_PATH |tr -d -c 0-9)
-            echo "[D] Parsed vars: CONFIG_PATH=$SURICATA_CONFIG_PATH TUN_NUMBER=$SURICATA_TUN_NUMBER"
-            nohup /usr/local/bin/suricata -i tun_$SURICATA_IF_NAME -D -c $SURICATA_CONFIG_PATH/suricata.yaml --pidfile /var/run/suricata_$SURICATA_IF_NAME$SURICATA_TUN_NUMBER.pid &
-
-            if ps aux | grep suricata | grep $SURICATA_IF_NAME 2> /dev/null
+            if ! ps -p $SURICATA_PID > /dev/null 2> /dev/null
             then
-                echo "[INFO] Suricata was successfully restarted on interface $SURICATA_IF_NAME!"
+                echo "[INFO] Suricata is NOT running on interface $SURICATA_IF_NAME! Will try to fix this..."
+                echo "[D] Removing potential stuck PID file"
+                rm "$VARPATH$SURICATA_IF_NAME"*
+
+                echo "[D] Restarting Suricata on interface"
+                SURICATA_CONFIG_PATH=$(find /usr/local/etc/suricata -path /usr/local/etc/suricata/suricata_*_$SURICATA_IF_NAME)
+                SURICATA_TUN_NUMBER=$(echo $SURICATA_CONFIG_PATH |tr -d -c 0-9)
+                echo "[D] Parsed vars: CONFIG_PATH=$SURICATA_CONFIG_PATH TUN_NUMBER=$SURICATA_TUN_NUMBER"
+                nohup /usr/local/bin/suricata -i tun_$SURICATA_IF_NAME -D -c $SURICATA_CONFIG_PATH/suricata.yaml --pidfile /var/run/suricata_$SURICATA_IF_NAME$SURICATA_TUN_NUMBER.pid &
+
+                if ps aux | grep suricata | grep $SURICATA_IF_NAME 2> /dev/null
+                then
+                    echo "[INFO] Suricata was successfully restarted on interface $SURICATA_IF_NAME!"
+                else
+                    echo "[WARNING] Could not restart Suricata on interface $SURICATA_IF_NAME"
+                fi
             else
-                echo "[WARNING] Could not restart Suricata on interface $SURICATA_IF_NAME"
+                echo "[INFO] Suricata is running on interface $SURICATA_IF_NAME."           
             fi
-        else
-            echo "[INFO] Suricata is running on interface $SURICATA_IF_NAME."           
         fi
     else
         echo "[WARNING] Did not found PID file for this interface. Check if interface name represents a correct interface configured in Suricata."
